+++ src/Helpers/UtilsGeneralHelper.php
@@ -347,7 +347,6 @@ final class UtilsGeneralHelper
 			'label' => '',
 			'icon' => '',
 			'type' => '',
-			'typeFilter' => '',
 			'itemId' => '',
 			'innerId' => '',
 			'fields' => [],
@@ -387,7 +386,6 @@ final class UtilsGeneralHelper
 		$settings = \apply_filters(UtilsConfig::FILTER_SETTINGS_DATA, [])[$type] ?? [];
 
 		$output['type'] = $type;
-		$output['typeFilter'] = $blockName['name'];
 		$output['integrationType'] = $settings['integrationType'] ?? '';
 		$output['label'] = $settings['labels']['title'] ?? '';
 		$output['icon'] = $settings['labels']['icon'] ?? '';
@@ -649,64 +647,13 @@ final class UtilsGeneralHelper
 	}
 
 	/**
-	 * Return files from data folder.
-	 *
-	 * @param string $type Folder name.
-	 * @param string $file File name with ext.
-	 *
-	 * @return array<mixed>
-	 */
-	public static function getDataManifest(string $type, string $file = 'manifest.json'): array
-	{
-		$path = self::getDataManifestRaw($type, $file);
-
-		if ($path) {
-			return \json_decode($path, true);
-		}
-
-		return [];
-	}
-
-	/**
-	 * Return files from data folder in raw format.
-	 *
-	 * @param string $type Folder name.
-	 * @param string $file File name with ext.
-	 *
-	 * @return string
-	 */
-	public static function getDataManifestRaw(string $type, string $file = 'manifest.json'): string
-	{
-		$path = self::getDataManifestPath($type, $file);
-
-		if (\file_exists($path)) {
-			return \file_get_contents($path); // phpcs:ignore WordPress.WP.AlternativeFunctions.file_get_contents_file_get_contents
-		}
-
-		return '';
-	}
-
-	/**
-	 * Return files full path.
-	 *
-	 * @param string $type Folder name.
-	 * @param string $file File name with ext.
-	 *
-	 * @return string
-	 */
-	public static function getDataManifestPath(string $type, string $file = 'manifest.json'): string
-	{
-		return \dirname(__FILE__, 3) . "/data/{$type}/{$file}";
-	}
-
-	/**
 	 * Return counries filtered by some key for multiple usages.
 	 *
 	 * @return array<int, array<int, string>>
 	 */
 	public static function getCountrySelectList(): array
 	{
-		return self::getDataManifest('country');
+		return UtilsDataHelper::getDataManifest('country');
 	}
 
 	/**
